Initial commit.
pub_classes.dart is a random sample of single-constructor classes from a
corpus of pub packages.

stateless_widgets.dart is a random sample of classes that extends
StatelessWidget from a corpus of open source code.

This commit is the original code with any clean-ups that could be
applied today: mostly using super parameters and getting rid of
unneeded super() clauses.