Require field parameters to be explicitly marked "var" or "final".
In the examples here, every parameter is a field parameter, so this is
mostly just adding "final" to the parameters in const constructors and
"var" to the parameters in the non-const constructors.

I didn't look around to see if there are classes that would benefit
from being able to access non-field parameters in a primary constructor.